+++ packages/ui/src/index.ts
@@ -44,3 +44,8 @@ export * from './Menu';
 export * from './Pagination';
 export * from './Popover';
 export * from './Stepper';
+
+export * from './Box';
+export * from './Container';
+export * from './Grid';
+export * from './Stack';